Client refactor, ClientException
