+++ client/src/main/java/org/folio/shared/index/client/ClientException.java
@@ -1,0 +1,12 @@
+package org.folio.shared.index.client;
+
+public class ClientException extends RuntimeException {
+
+  ClientException(String s) {
+    super(s);
+  }
+
+  ClientException(Throwable t) {
+    super(t);
+  }
+}
